using double in extracting from json variable

+++ src/model/process_model.h
@@ -51,9 +51,9 @@ class ProcessModel : public BaseModel {
     MODEL_PROP(int, printPercentage, 0)
     MODEL_PROP(int, timeRemaining, 0)
     MODEL_PROP(int, errorCode, 0)
-    MODEL_PROP(float, targetHesUpper, 0)
-    MODEL_PROP(float, targetHesLower, 0)
-    MODEL_PROP(float, currentHes, 0)
+    MODEL_PROP(double, targetHesUpper, 0)
+    MODEL_PROP(double, targetHesLower, 0)
+    MODEL_PROP(double, currentHes, 0)
     MODEL_PROP(int, levelState, 0)
 
   public:
